assignment-1-101699682/code/requirements.txt
pandas==2.1.4
numpy==1.26.4
cassandra-driver==3.29.0